fsec/hostdiscovery/requirements.txt
alabaster==0.7.12
Babel==2.9.1
certifi==2021.5.30
charset-normalizer==2.0.3
docutils==0.16
grpcio==1.39.0
grpcio-tools==1.39.0
idna==3.2
imagesize==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
protobuf==3.17.3
Pygments==2.9.0
pyparsing==2.4.7
python3-nmap==1.5.0
pytz==2021.1
requests==2.26.0
simplejson==3.17.3
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.1.1
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
urllib3==1.26.6
